+++ src/IPFees.Calculator/Evaluator/NodeBinary.cs
@@ -7,7 +7,7 @@ namespace IPFees.Evaluator
     {
         // Constructor accepts the two nodes to be operated on and function
         // that performs the actual operation
-        public NodeBinary(Node lhs, Node rhs, Func<double, double, double> op)
+        public NodeBinary(Node lhs, Node rhs, Func<decimal, decimal, decimal> op)
         {
             _lhs = lhs;
             _rhs = rhs;
@@ -16,9 +16,9 @@ namespace IPFees.Evaluator
 
         Node _lhs;                              // Left hand side of the operation
         Node _rhs;                              // Right hand side of the operation
-        Func<double, double, double> _op;       // The callback operator
+        Func<decimal, decimal, decimal> _op;       // The callback operator
 
-        public override double Eval(IContext ctx)
+        public override decimal Eval(IContext ctx)
         {
             // Evaluate both sides
             var lhsVal = _lhs.Eval(ctx);
